+++ src/test/encrypt/encrypt-scheme_BFV-poly_16384-sec_256-type_int32.test.js
@@ -11,12 +11,11 @@ describe('encrypt on BFV', () => {
         polyModulusDegree: 16384
       })
 
-      // Create a suitable vector of CoeffModulus primes
+      // Create a suitable set of CoeffModulus primes
       parms.setCoeffModulus({
         coeffModulus: Morfix.CoeffModulus.Create({
           polyModulusDegree: 16384,
-          bitSizes: Morfix.Vector({array: new Int32Array([47,47,47,48,48]) }),
-          securityLevel: Morfix.SecurityLevel.tc256
+          bitSizes: Int32Array.from([47,47,47,48,48])
         })
       })
 
@@ -60,46 +59,36 @@ describe('encrypt on BFV', () => {
         length: 16384
       }).map((x, i) =>  i)
 
-      // Convert data to a c++ 'vector'
-      const vector = Morfix.Vector({array})
-
       // Create a plainText variable and encode the vector to it
       const plainText = Morfix.PlainText()
 
-      encoder.encodeVectorInt32({
-        vector: vector,
-        plainText: plainText
+      encoder.encode({
+        array,
+        plainText
       })
 
       // Create a cipherText variable and encrypt the plainText to it
       const cipherText = Morfix.CipherText()
       encryptor.encrypt({
-        plainText: plainText,
-        cipherText: cipherText
+        plainText,
+        cipherText
       })
 
       // Create a new plainText variable to store the decrypted cipherText
       const decryptedPlainText = Morfix.PlainText()
       decryptor.decrypt({
-        cipherText: cipherText,
+        cipherText,
         plainText: decryptedPlainText
       })
 
-      // Create a c++ vector to store the decoded result
-      const decodeVector = Morfix.Vector({array: new Int32Array() })
-
-      // Decode the PlainText to the c++ vector
-      encoder.decodeVectorInt32({
-        plainText: decryptedPlainText,
-        vector: decodeVector
+      // Decode the PlainText
+      const decodedArray = encoder.decode({
+        plainText: decryptedPlainText
       })
 
-      // Convert the vector to a JS array
-      const decryptedArray = decodeVector.toArray()
-
-      expect(decryptedArray).toBeInstanceOf(Int32Array)
+      expect(decodedArray).toBeInstanceOf(Int32Array)
       // Check values
-      expect(decryptedArray).toEqual(array)
+      expect(decodedArray).toEqual(array)
     })
   })
 })