+++ src/test/galoiskey/galoiskey-scheme_BFV-poly_16384-sec_128.test.js
@@ -10,16 +10,15 @@ describe.skip('galoiskey on BFV', () => {
       parms.setPolyModulusDegree({
         polyModulusDegree: 16384
       })
-        
-      // Create a suitable vector of CoeffModulus primes
+
+      // Create a suitable set of CoeffModulus primes
       parms.setCoeffModulus({
         coeffModulus: Morfix.CoeffModulus.Create({
           polyModulusDegree: 16384,
-          bitSizes: Morfix.Vector({array: new Int32Array([48,48,48,49,49,49,49,49,49]) }),
-          securityLevel: Morfix.SecurityLevel.tc128
+          bitSizes: Int32Array.from([48,48,48,49,49,49,49,49,49])
         })
       })
-      
+
       // Set the PlainModulus to a prime of bitSize 20.
       parms.setPlainModulus({
         plainModulus: Morfix.PlainModulus.Batching({
@@ -27,7 +26,7 @@ describe.skip('galoiskey on BFV', () => {
           bitSize: 20
         })
       })
-      
+
       const context = Morfix.Context({
         encryptionParams: parms,
         expandModChain: true,
@@ -39,15 +38,15 @@ describe.skip('galoiskey on BFV', () => {
       const keyGenerator = Morfix.KeyGenerator({
         context: context
       })
-      
+
       const spyGenGaloisKeys = jest.spyOn(keyGenerator, 'genGaloisKeys')
       const galoisKeys = keyGenerator.genGaloisKeys()
       expect(spyGenGaloisKeys).toHaveBeenCalled()
-      
+
       const spySaveGaloisKeys = jest.spyOn(galoisKeys, 'save')
       const base64 = galoisKeys.save()
       expect(spySaveGaloisKeys).toHaveBeenCalled()
-      
+
       const spyLoadGaloisKeys = jest.spyOn(galoisKeys, 'load')
       galoisKeys.load({context, encoded: base64})
       expect(spyLoadGaloisKeys).toHaveBeenCalled()