+++ src/test/generateKeyPairTests.js
@@ -39,19 +39,18 @@ const genTests = (verb) => {
       parms.setPolyModulusDegree({
         polyModulusDegree: ${POLYMODULUS_DEGREES[polyModDeg]}
       })
-        `)
+`)
 
-          if (schemeType === SCHEME_TYPES.BFV) {
-            code.push(`
-      // Create a suitable vector of CoeffModulus primes
+        if (schemeType === SCHEME_TYPES.BFV) {
+          code.push(`
+      // Create a suitable set of CoeffModulus primes
       parms.setCoeffModulus({
         coeffModulus: Morfix.CoeffModulus.Create({
           polyModulusDegree: ${POLYMODULUS_DEGREES[polyModDeg]},
-          bitSizes: Morfix.Vector({array: new Int32Array([${BFV_COEFF_MOD_BIT_SIZES[SECURITY_LEVELS[secLevel]][POLYMODULUS_DEGREES[polyModDeg]]}]) }),
-          securityLevel: ${SECURITY_LEVELS_CONSTRUCTOR[SECURITY_LEVELS[secLevel]]}
+          bitSizes: Int32Array.from([${BFV_COEFF_MOD_BIT_SIZES[SECURITY_LEVELS[secLevel]][POLYMODULUS_DEGREES[polyModDeg]]}])
         })
       })
-      
+
       // Set the PlainModulus to a prime of bitSize 20.
       parms.setPlainModulus({
         plainModulus: Morfix.PlainModulus.Batching({
@@ -59,21 +58,20 @@ const genTests = (verb) => {
           bitSize: 20
         })
       })
-      `)
-          }
+`)
+        }
 
-          if (schemeType === SCHEME_TYPES.CKKS) {
-            code.push(`
-      // Create a suitable vector of CoeffModulus primes (we use default set)
+        if (schemeType === SCHEME_TYPES.CKKS) {
+          code.push(`
+      // Create a suitable set of CoeffModulus primes (we use default set)
       parms.setCoeffModulus({
         coeffModulus: Morfix.CoeffModulus.Create({
           polyModulusDegree: ${POLYMODULUS_DEGREES[polyModDeg]},
-          bitSizes: Morfix.Vector({array: new Int32Array([${CKKS_COEFF_MOD_BIT_SIZES[SECURITY_LEVELS[secLevel]][POLYMODULUS_DEGREES[polyModDeg]]}]) }),
-          securityLevel: ${SECURITY_LEVELS_CONSTRUCTOR[SECURITY_LEVELS[secLevel]]}
+          bitSizes: Int32Array.from([${CKKS_COEFF_MOD_BIT_SIZES[SECURITY_LEVELS[secLevel]][POLYMODULUS_DEGREES[polyModDeg]]}])
         })
       })
-      `)
-          }
+`)
+        }
 
 
           code.push(`
@@ -88,20 +86,20 @@ const genTests = (verb) => {
       const keyGenerator = Morfix.KeyGenerator({
         context: context
       })
-      
+
       const spyGetSecretKey = jest.spyOn(keyGenerator, 'getSecretKey')
       const secretKey = keyGenerator.getSecretKey()
       expect(spyGetSecretKey).toHaveBeenCalled()
-      
+
       const spyGetPublicKey = jest.spyOn(keyGenerator, 'getPublicKey')
       const publicKey = keyGenerator.getPublicKey()
       expect(spyGetPublicKey).toHaveBeenCalled()
 
-      
+
       const spySaveSecretKey = jest.spyOn(secretKey, 'save')
       const secretKeyBase64 = secretKey.save()
       expect(spySaveSecretKey).toHaveBeenCalled()
-      
+
       const spySavePublicKey = jest.spyOn(publicKey, 'save')
       const publicKeyBase64 = publicKey.save()
       expect(spySavePublicKey).toHaveBeenCalled()
@@ -110,13 +108,14 @@ const genTests = (verb) => {
       const spyLoadSecretKey = jest.spyOn(secretKey, 'load')
       secretKey.load({context, encoded: secretKeyBase64})
       expect(spyLoadSecretKey).toHaveBeenCalled()
-      
+
       const spyLoadPublicKey = jest.spyOn(publicKey, 'load')
       publicKey.load({context, encoded: publicKeyBase64})
       expect(spyLoadPublicKey).toHaveBeenCalled()
     })
   })
-})`)
+})
+`)
           fs.mkdirSync(`${process.cwd()}/${folderName}/`, {recursive: true})
           fs.writeFileSync(`./${folderName}/${fileName}`, code.join(''))
       }