+++ CHANGES.md
@@ -3,6 +3,20 @@
 See [Microsoft's Change log](https://github.com/microsoft/SEAL/blob/master/Changes.md)
 for more details on each SEAL version change.
 
+## Version 3.2.0
+
+Feat:
+- No more dealing with `Vector` types! All methods now accept a valid 
+ TypedArray as a parameter or return one where applicable
+
+Deprecated:
+- In `BatchEncode`, replace `encodeVectorInt32`, `encodeVectorUInt32` with `BatchEncode.encode` using a TypedArray
+- In `BatchEncode`, replace `decodeVectorInt32`, `decodeVectorUInt32` with `BatchEncode.decode`. This method returns
+  a either an Int32Array or Uint32Array with the results. 
+- In `CKKSEncode`, replace `decodeVectorDouble` with `CKKSEncode.decode`. This method returns
+  a Float64Array with the results.
+- `CoeffModulus.Create` now accepts `bitSizes` as an Int32Array in addition to the legacy Vector type.
+
 ## Version 3.1.0
 
 Fix: